Fix numbers that have decimals when displaying in grid/summary.

+++ assets/opstools/AppBuilder/components/ab_work_object_workspace_datatable.js
@@ -689,7 +689,7 @@ module.exports = class ABWorkObjectDatatable extends ABComponent {
             if (editor.config) {
                switch (editor.config.editor) {
                   case "number":
-                     state.value = parseInt(state.value);
+                     state.value = parseFloat(state.value);
                      break;
                   case "datetime":
                      state.value = state.value.getTime();
